+++ static/app/views/dashboardsV2/widgetCard/widgetCardChartContainer.tsx
@@ -1,37 +1,21 @@
-import * as React from 'react';
+import {Fragment} from 'react';
 import {withRouter, WithRouterProps} from 'react-router';
 import styled from '@emotion/styled';
 import {LegendComponentOption} from 'echarts';
 
 import {Client} from 'sentry/api';
-import ErrorPanel from 'sentry/components/charts/errorPanel';
-import SimpleTableChart from 'sentry/components/charts/simpleTableChart';
 import TransparentLoadingMask from 'sentry/components/charts/transparentLoadingMask';
 import LoadingIndicator from 'sentry/components/loadingIndicator';
-import Placeholder from 'sentry/components/placeholder';
-import {IconWarning} from 'sentry/icons';
-import space from 'sentry/styles/space';
 import {Organization, PageFilters} from 'sentry/types';
 import {EChartDataZoomHandler, EChartEventHandler, Series} from 'sentry/types/echarts';
-import {defined} from 'sentry/utils';
-import {getIssueFieldRenderer} from 'sentry/utils/dashboards/issueFieldRenderers';
-import {TableDataRow} from 'sentry/utils/discover/discoverQuery';
-import {eventViewFromWidget} from 'sentry/views/dashboardsV2/utils';
 
 import {Widget, WidgetType} from '../types';
-import {
-  ISSUE_FIELD_TO_HEADER_MAP,
-  ISSUE_FIELDS,
-} from '../widgetBuilder/issueWidget/fields';
 
 import WidgetCardChart from './chart';
+import {IssueWidgetCard} from './issueWidgetCard';
 import IssueWidgetQueries from './issueWidgetQueries';
 import MetricsWidgetQueries from './metricsWidgetQueries';
 import WidgetQueries from './widgetQueries';
-
-type TableResultProps = Pick<WidgetQueries['state'], 'errorMessage' | 'loading'> & {
-  transformedResults: TableDataRow[];
-};
 
 type Props = WithRouterProps & {
   api: Client;
@@ -70,55 +54,7 @@ export function WidgetCardChartContainer({
   expandNumbers,
   onDataFetched,
 }: Props) {
-  function issueTableResultComponent({
-    loading,
-    errorMessage,
-    transformedResults,
-  }: TableResultProps): React.ReactNode {
-    if (errorMessage) {
-      return (
-        <ErrorPanel>
-          <IconWarning color="gray500" size="lg" />
-        </ErrorPanel>
-      );
-    }
-
-    if (loading) {
-      // Align height to other charts.
-      return <LoadingPlaceholder height="200px" />;
-    }
-
-    const query = widget.queries[0];
-    const queryFields = defined(query.fields)
-      ? query.fields
-      : [...query.columns, ...query.aggregates];
-    const fieldAliases = query.fieldAliases ?? [];
-    const eventView = eventViewFromWidget(
-      widget.title,
-      widget.queries[0],
-      selection,
-      widget.displayType
-    );
-
-    return (
-      <StyledSimpleTableChart
-        location={location}
-        title=""
-        eventView={eventView}
-        fields={queryFields}
-        fieldAliases={fieldAliases}
-        loading={loading}
-        metadata={ISSUE_FIELDS}
-        data={transformedResults}
-        organization={organization}
-        getCustomFieldRenderer={getIssueFieldRenderer}
-        fieldHeaderMap={ISSUE_FIELD_TO_HEADER_MAP}
-        stickyHeaders
-      />
-    );
-  }
-
-  function renderIssueChart() {
+  if (widget.widgetType === WidgetType.ISSUE) {
     return (
       <IssueWidgetQueries
         api={api}
@@ -129,24 +65,28 @@ export function WidgetCardChartContainer({
       >
         {({transformedResults, errorMessage, loading}) => {
           return (
-            <React.Fragment>
+            <Fragment>
               {typeof renderErrorMessage === 'function'
                 ? renderErrorMessage(errorMessage)
                 : null}
               <LoadingScreen loading={loading} />
-              {issueTableResultComponent({
-                transformedResults,
-                loading,
-                errorMessage,
-              })}
-            </React.Fragment>
+              <IssueWidgetCard
+                transformedResults={transformedResults}
+                loading={loading}
+                errorMessage={errorMessage}
+                widget={widget}
+                organization={organization}
+                location={location}
+                selection={selection}
+              />
+            </Fragment>
           );
         }}
       </IssueWidgetQueries>
     );
   }
 
-  function renderMetricsChart() {
+  if (widget.widgetType === WidgetType.METRICS) {
     return (
       <MetricsWidgetQueries
         api={api}
@@ -157,7 +97,7 @@ export function WidgetCardChartContainer({
       >
         {({tableResults, timeseriesResults, errorMessage, loading}) => {
           return (
-            <React.Fragment>
+            <Fragment>
               {typeof renderErrorMessage === 'function'
                 ? renderErrorMessage(errorMessage)
                 : null}
@@ -174,62 +114,50 @@ export function WidgetCardChartContainer({
                 isMobile={isMobile}
                 windowWidth={windowWidth}
               />
-            </React.Fragment>
+            </Fragment>
           );
         }}
       </MetricsWidgetQueries>
     );
   }
 
-  function renderDiscoverChart() {
-    return (
-      <WidgetQueries
-        api={api}
-        organization={organization}
-        widget={widget}
-        selection={selection}
-        limit={tableItemLimit}
-        onDataFetched={onDataFetched}
-      >
-        {({tableResults, timeseriesResults, errorMessage, loading}) => {
-          return (
-            <React.Fragment>
-              {typeof renderErrorMessage === 'function'
-                ? renderErrorMessage(errorMessage)
-                : null}
-              <WidgetCardChart
-                timeseriesResults={timeseriesResults}
-                tableResults={tableResults}
-                errorMessage={errorMessage}
-                loading={loading}
-                location={location}
-                widget={widget}
-                selection={selection}
-                router={router}
-                organization={organization}
-                isMobile={isMobile}
-                windowWidth={windowWidth}
-                onZoom={onZoom}
-                onLegendSelectChanged={onLegendSelectChanged}
-                legendOptions={legendOptions}
-                expandNumbers={expandNumbers}
-              />
-            </React.Fragment>
-          );
-        }}
-      </WidgetQueries>
-    );
-  }
-
-  if (widget.widgetType === WidgetType.ISSUE) {
-    return renderIssueChart();
-  }
-
-  if (widget.widgetType === WidgetType.METRICS) {
-    return renderMetricsChart();
-  }
-
-  return renderDiscoverChart();
+  return (
+    <WidgetQueries
+      api={api}
+      organization={organization}
+      widget={widget}
+      selection={selection}
+      limit={tableItemLimit}
+      onDataFetched={onDataFetched}
+    >
+      {({tableResults, timeseriesResults, errorMessage, loading}) => {
+        return (
+          <Fragment>
+            {typeof renderErrorMessage === 'function'
+              ? renderErrorMessage(errorMessage)
+              : null}
+            <WidgetCardChart
+              timeseriesResults={timeseriesResults}
+              tableResults={tableResults}
+              errorMessage={errorMessage}
+              loading={loading}
+              location={location}
+              widget={widget}
+              selection={selection}
+              router={router}
+              organization={organization}
+              isMobile={isMobile}
+              windowWidth={windowWidth}
+              onZoom={onZoom}
+              onLegendSelectChanged={onLegendSelectChanged}
+              legendOptions={legendOptions}
+              expandNumbers={expandNumbers}
+            />
+          </Fragment>
+        );
+      }}
+    </WidgetQueries>
+  );
 }
 
 export default withRouter(WidgetCardChartContainer);
@@ -252,15 +180,3 @@ const LoadingScreen = ({loading}: {loading: boolean}) => {
     </StyledTransparentLoadingMask>
   );
 };
-
-const LoadingPlaceholder = styled(Placeholder)`
-  background-color: ${p => p.theme.surface200};
-`;
-
-const StyledSimpleTableChart = styled(SimpleTableChart)`
-  margin-top: ${space(1.5)};
-  border-bottom-left-radius: ${p => p.theme.borderRadius};
-  border-bottom-right-radius: ${p => p.theme.borderRadius};
-  font-size: ${p => p.theme.fontSizeMedium};
-  box-shadow: none;
-`;
