+++ static/app/views/dashboardsV2/widgetCard/widgetQueries.tsx
@@ -221,7 +221,6 @@ class WidgetQueries extends React.Component<Props, State> {
       !isSelectionEqual(selection, prevProps.selection) ||
       cursor !== prevProps.cursor
     ) {
-      console.log('PRISCILA');
       this.fetchData();
       return;
     }
