+++ tests/js/spec/views/dashboardsV2/widgetBuilder/widgetBuilder.spec.tsx
@@ -29,6 +29,8 @@ const defaultOrgFeatures = [
   'global-views',
 ];
 
+jest.unmock('lodash/debounce');
+
 // Mocking worldMapChart to avoid act warnings
 jest.mock('sentry/components/charts/worldMapChart');
 
@@ -110,6 +112,7 @@ describe('WidgetBuilder', function () {
 
   let eventsStatsMock: jest.Mock | undefined;
   let eventsv2Mock: jest.Mock | undefined;
+  let issuesMock: jest.Mock | undefined;
   let metricsDataMock: jest.Mock | undefined;
   let tagsMock: jest.Mock | undefined;
 
@@ -157,8 +160,9 @@ describe('WidgetBuilder', function () {
       body: [],
     });
 
-    MockApiClient.addMockResponse({
+    issuesMock = MockApiClient.addMockResponse({
       url: '/organizations/org-slug/issues/',
+      method: 'GET',
       body: [],
     });
 
@@ -1609,6 +1613,7 @@ describe('WidgetBuilder', function () {
   // Disabling for CI, but should run locally when making changes
   // eslint-disable-next-line jest/no-disabled-tests
   it.skip('Update table header values (field alias)', async function () {
+    jest.useFakeTimers();
     const handleSave = jest.fn();
 
     renderTestComponent({
@@ -1619,10 +1624,18 @@ describe('WidgetBuilder', function () {
     await screen.findByText('Table');
 
     userEvent.type(screen.getByPlaceholderText('Alias'), 'First Alias{enter}');
+    act(() => {
+      jest.advanceTimersByTime(DEFAULT_DEBOUNCE_DURATION + 1);
+    });
 
     userEvent.click(screen.getByLabelText('Add a Column'));
 
     userEvent.type(screen.getAllByPlaceholderText('Alias')[1], 'Second Alias{enter}');
+    act(() => {
+      jest.advanceTimersByTime(DEFAULT_DEBOUNCE_DURATION + 1);
+    });
+
+    expect(eventsv2Mock).toHaveBeenCalledTimes(3);
 
     userEvent.click(screen.getByText('Add Widget'));
 
@@ -1637,7 +1650,7 @@ describe('WidgetBuilder', function () {
     });
   });
 
-  describe.only('Issue Widgets', function () {
+  describe('Issue Widgets', function () {
     it('sets widgetType to issues', async function () {
       const handleSave = jest.fn();
 
@@ -1739,7 +1752,7 @@ describe('WidgetBuilder', function () {
 
     // Disabling for CI, but should run locally when making changes
     // eslint-disable-next-line jest/no-disabled-tests
-    it.only('Update table header values (field alias)', async function () {
+    it.skip('Update table header values (field alias)', async function () {
       jest.useFakeTimers();
       const handleSave = jest.fn();
 
@@ -1750,18 +1763,26 @@ describe('WidgetBuilder', function () {
 
       await screen.findByText('Table');
 
+      expect(eventsv2Mock).toHaveBeenCalledTimes(1);
+
       userEvent.click(screen.getByText('Issues (Status, assignee, etc.)'));
 
-      await screen.findAllByPlaceholderText('Alias');
-
       userEvent.type(screen.getAllByPlaceholderText('Alias')[0], 'First Alias{enter}');
-      jest.advanceTimersByTime(DEFAULT_DEBOUNCE_DURATION + 1);
+      act(() => {
+        jest.advanceTimersByTime(DEFAULT_DEBOUNCE_DURATION + 1);
+      });
 
       userEvent.type(screen.getAllByPlaceholderText('Alias')[1], 'Second Alias{enter}');
-      jest.advanceTimersByTime(DEFAULT_DEBOUNCE_DURATION + 1);
+      act(() => {
+        jest.advanceTimersByTime(DEFAULT_DEBOUNCE_DURATION + 1);
+      });
 
       userEvent.type(screen.getAllByPlaceholderText('Alias')[2], 'Third Alias{enter}');
-      jest.advanceTimersByTime(DEFAULT_DEBOUNCE_DURATION + 1);
+      act(() => {
+        jest.advanceTimersByTime(DEFAULT_DEBOUNCE_DURATION + 1);
+      });
+
+      expect(issuesMock).toHaveBeenCalledTimes(3);
 
       userEvent.click(screen.getByText('Add Widget'));
